tests/requirements.txt
# =========================
# Pytest & test utilities
# =========================
pytest==8.3.3
pytest-env==1.1.3
requests==2.32.3

# =========================
# FastAPI stack
# =========================
fastapi==0.115.6
starlette==0.41.3
httpx==0.27.2

# =========================
# Data validation / config
# =========================
pydantic==2.10.4
pydantic-settings==2.6.1
python-dotenv==1.0.1

# =========================
# Database / ORM
# =========================
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pgvector==0.3.6

# =========================
# Redis
# =========================
redis==5.2.1
rq==1.16.2

# =========================
# S3 / MinIO client
# =========================
boto3==1.35.67
botocore==1.35.67

# =========================
# Files
# =========================
PyYAML==6.0.2
